Got rid of some unused local variables, cleaned up the imports, etc...  No major code changes.

+++ Evolvo/src/org/maloi/evolvo/expressiontree/operators/scalar/Mandel.java
@@ -84,7 +84,7 @@ public class Mandel implements OperatorInterface, Serializable
       }
       while ((lengthz <= 4.0) && (Iteration < 255));
 
-      theStack.push((double) Iteration / 128.0 - 1.0);
+      theStack.push(Iteration / 128.0 - 1.0);
    }
 
    /** Returns the operator's name. */
